app/requirements.txt
aiogram==2.11.2
aiohttp==3.7.2
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.11.8
chardet==3.0.4
EasyProcess==0.3
hiredis==1.1.0
idna==2.10
iniconfig==1.1.1
MarkupSafe==1.1.1
multidict==5.0.0
packaging==20.4
pluggy==0.13.1
prodict==0.8.11
py==1.9.0
pyaml==20.4.0
pyparsing==2.4.7
pytest==6.1.2
pytest-mock==3.3.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.1
yarl==1.6.2
tqdm
